+++ examples/_extensions/kbodwin/flourish/flourish.lua
@@ -1,0 +1,5 @@
+quarto.doc.add_html_dependency({
+  name = "flourish",
+  version = "0.0.1",
+  scripts = { "injectFlourishes.js", "flourish.js" }
+})